flask_api/requirements.txt
pandas==2.2.3
numpy==1.26.4
tensorflow==2.18.0
scikit-learn==1.6.0
flask==3.1.0
Pillow==11.0.0
requests==2.32.3
streamlit==1.41.1

